error on creating existing file

+++ src/utils/createFile.ts
@@ -12,8 +12,15 @@ export const createFile = async (
 	const newFilePath = path.join(fileConnectionDir, newFileName)
 	const fileContent = `// CodeGraphy connect: './${connectionName}'`
 
-	fs.writeFileSync(newFilePath, fileContent)
-	const newFileUri = vscode.Uri.file(newFilePath)
-	const document = await vscode.workspace.openTextDocument(newFileUri)
-	await vscode.window.showTextDocument(document)
+	// Only create the file if it doesn't exist
+	if (!fs.existsSync(newFilePath)) {
+		fs.writeFileSync(newFilePath, fileContent)
+		const newFileUri = vscode.Uri.file(newFilePath)
+		const document = await vscode.workspace.openTextDocument(newFileUri)
+		await vscode.window.showTextDocument(document)
+	} else {
+		vscode.window.showErrorMessage(
+			`File ${newFileName} already exists in ${fileConnectionDir}`,
+		)
+	}
 }
